:bug: Fix: XML 파싱에 따른 모든 응답 API XML 형식 변경 문제 해결 (#70)
